Checkpoints and Death from falls added
Checkpoints have been implemented just need to add animation to change the sprite.
Added a hazard layer that kills the player when they fall down the holes in the scenes and respawns the player at the last checkpoint.
Player currently has 3 lives.
If all lives lost then it resets the level.